changed interface location and renamed pagename
