+++ server/src/shopLists/coreShops.ts
@@ -1,12 +1,12 @@
-import { IShopLocation } from "../interfaces/IShop";
+import { IShopLocation } from '../../../shared/interfaces';
 
 export const coreShopLocations: IShopLocation[] = [
-    { x: 25.980966567993164, y: -1345.6417236328125, z: 28.497024536132812, isBlip: true }, 
-    { x: 374.3475341796875, y: 328.112060546875, z: 102.56637573242188, isBlip: true }, 
-    { x: -3041.32763671875, y: 585.155029296875, z: 6.908928871154785, isBlip: true }, 
-    { x: -3243.743408203125, y: 1001.3903198242188, z: 11.830706596374512, isBlip: true }, 
-    { x: 548.0447387695312, y: 2669.48876953125, z: 41.156490325927734, isBlip: true }, 
-    { x: 1960.2322998046875, y: 3742.317138671875, z: 31.343746185302734, isBlip: true }, 
-    { x: 1730.01171875, y: 6416.22021484375, z: 34.03722381591797, isBlip: true }, 
-    { x: 2555.4609375, y: 382.1643371582031, z: 107.62295532226562, isBlip: true }, 
+    { x: 25.980966567993164, y: -1345.6417236328125, z: 28.497024536132812, isBlip: true },
+    { x: 374.3475341796875, y: 328.112060546875, z: 102.56637573242188, isBlip: true },
+    { x: -3041.32763671875, y: 585.155029296875, z: 6.908928871154785, isBlip: true },
+    { x: -3243.743408203125, y: 1001.3903198242188, z: 11.830706596374512, isBlip: true },
+    { x: 548.0447387695312, y: 2669.48876953125, z: 41.156490325927734, isBlip: true },
+    { x: 1960.2322998046875, y: 3742.317138671875, z: 31.343746185302734, isBlip: true },
+    { x: 1730.01171875, y: 6416.22021484375, z: 34.03722381591797, isBlip: true },
+    { x: 2555.4609375, y: 382.1643371582031, z: 107.62295532226562, isBlip: true },
 ];
