+++ server/src/shopLists/juice.ts
@@ -1,4 +1,4 @@
-import { IShopLocation } from '../interfaces/IShop';
+import { IShopLocation } from '../../../shared/interfaces';
 
 export const juiceLocations: IShopLocation[] = [
     { x: 1166.243896484375, y: 2709.356689453125, z: 37.15770721435547, isBlip: true },
