+++ server/src/shopLists/liquorAce.ts
@@ -1,4 +1,4 @@
-import { IShopLocation } from '../interfaces/IShop';
+import { IShopLocation } from '../../../shared/interfaces';
 
 export const liquorAceLocations: IShopLocation[] = [
     { x: 1393.5035400390625, y: 3605.268798828125, z: 33.98093032836914, isBlip: true },
