+++ server/src/shopLists/ltd.ts
@@ -1,9 +1,9 @@
-import { IShopLocation } from '../interfaces/IShop';
+import { IShopLocation } from '../../../shared/interfaces';
 
 export const ltdLocations: IShopLocation[] = [
     { x: -48.5690803527832, y: -1757.6961669921875, z: 28.4210147857666, isBlip: true },
     { x: 1163.400634765625, y: -323.938232421875, z: 68.20509338378906, isBlip: true },
     { x: -707.4390258789062, y: -914.5612182617188, z: 18.21558952331543, isBlip: true },
     { x: 1697.968505859375, y: 4924.54833984375, z: 41.06367492675781, isBlip: true },
-    { x: -1820.6717529296875, y: 792.8975219726562, z: 137.11163330078125, isBlip: true }, 
+    { x: -1820.6717529296875, y: 792.8975219726562, z: 137.11163330078125, isBlip: true },
 ];
