+++ server/src/shopLists/robsLiquor.ts
@@ -1,8 +1,8 @@
-import { IShopLocation } from '../interfaces/IShop';
+import { IShopLocation } from '../../../shared/interfaces';
 
 export const robsLiquorLocations: IShopLocation[] = [
     { x: 1135.9544677734375, y: -981.8599853515625, z: 45.41580581665039, isBlip: true },
     { x: -1222.91162109375, y: -907.1942749023438, z: 11.326356887817383, isBlip: true },
     { x: -1487.08349609375, y: -379.2518005371094, z: 39.163429260253906, isBlip: true },
-    { x: -2967.888427734375, y: 390.76654052734375, z: 14.043313026428223, isBlip: true }, 
+    { x: -2967.888427734375, y: 390.76654052734375, z: 14.043313026428223, isBlip: true },
 ];
