+++ server/src/shopLists/sellerExample.ts
@@ -1,4 +1,4 @@
-import { IShopLocation } from '../interfaces/IShop';
+import { IShopLocation } from '../../../shared/interfaces';
 
 export const sellerExampleLocations: IShopLocation[] = [
     { x: 1163.400634765625, y: -323.938232421875, z: 68.20509338378906, isBlip: true },
