+++ server/src/shopLists/tequiLaLa.ts
@@ -1,4 +1,4 @@
-import { IShopLocation } from '../interfaces/IShop';
+import { IShopLocation } from '../../../shared/interfaces';
 
 export const tequiLaLaLocations: IShopLocation[] = [
     { x: -560.1758422851562, y: 286.4307556152344, z: 82.17138671875, isBlip: true },
