+++ server/src/shopLists/toolShop.ts
@@ -1,4 +1,4 @@
-import { IShopLocation } from '../interfaces/IShop';
+import { IShopLocation } from '../../../shared/interfaces';
 
 export const toolShopLocations: IShopLocation[] = [
     { x: 2748.77392578125, y: 3472.442626953125, z: 55.67741012573242 - 1, isBlip: true },
