+++ server/src/shopLists/vanillaUnicorn.ts
@@ -1,4 +1,4 @@
-import { IShopLocation } from '../interfaces/IShop';
+import { IShopLocation } from '../../../shared/interfaces';
 
 export const vanillaUnicornLocations: IShopLocation[] = [
     { x: 127.16043853759766, y: -1283.4989013671875, z: 29.2630615234375, isBlip: true },
